--- src/edu/self/startux/lastLog/FirstLogExecutor.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2012 StarTux.
- *
- * This file is part of LastLog.
- *
- * LastLog is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * LastLog is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with LastLog.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package edu.self.startux.lastLog;
-
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandExecutor;
-import org.bukkit.command.CommandSender;
-
-public class FirstLogExecutor implements CommandExecutor {
-        private LastLogPlugin plugin;
-
-        public FirstLogExecutor(LastLogPlugin plugin) {
-                this.plugin = plugin;
-        }
-
-        @Override
-        public boolean onCommand(CommandSender sender, Command command, String label, String[] args) {
-                return plugin.executeSomeLog(sender, label, args, false);
-        }
-}
