--- src/edu/self/startux/lastLog/LastLogColors.java
@@ -1,34 +1,0 @@
-/* * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
- * Copyright 2012 StarTux.
- *
- * This file is part of LastLog.
- *
- * LastLog is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * LastLog is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with LastLog.  If not, see <http://www.gnu.org/licenses/>.
- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
-
-package edu.self.startux.lastLog;
-
-import org.bukkit.ChatColor;
-
-public class LastLogColors {
-        final public static String PLAYER_NAME = "" + ChatColor.GOLD;
-        final public static String HEADER = "" + ChatColor.AQUA;
-        final public static String MINOR = "" + ChatColor.DARK_AQUA;
-        final public static String DATE = "" + ChatColor.YELLOW;
-        final public static String UNKNOWN = "" + ChatColor.GRAY;
-        final public static String ERROR = "" + ChatColor.RED;
-        final public static String RESET = "" + ChatColor.WHITE;
-        final public static String ONLINE = "" + ChatColor.GREEN;
-        final public static String OFFLINE = "" + ChatColor.RED;
-}
